third_party/blink/web_tests/fast/dom/TreeWalker/traversal-skip-most-expected.txt

Test TreeWalker with skipping

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing nextSibling
PASS walker.firstChild(); walker.currentNode.id is 'B1'
PASS walker.nextSibling(); walker.currentNode.id is 'B3'

Testing previousSibling
PASS walker.previousSibling(); walker.currentNode.id is 'B1'
PASS successfullyParsed is true

TEST COMPLETE
